aperisolve/requirements.txt
Flask
Flask-SQLAlchemy
numpy
redis
rq
psycopg2-binary
types-Flask-SQLAlchemy
sqlalchemy-stubs
Pillow
gunicorn